[refactor] test code 병합
